+++ NEWS.md
@@ -2,13 +2,14 @@
 
 ## Significant interface changes
 
+* Allow authentication without having to create a service principal first, by leveraging the Azure CLI cross-platform app. It's still recommended to create your own SP for authentication, if possible.
 * New `create_azure_login`, `get_azure_login` and `delete_azure_login` functions to handle ARM authentication. While directly calling `az_rm$new()` will still work, it's recommended to use `create_azure_login` and `get_azure_login` going forward. Login credentials will be saved and reused for subsequent sessions (see below).
 * `get_azure_token` significantly revamped. It now supports four authentication methods for obtaining AAD tokens:
   - Client credentials (what you would use with a "web app" registered service principal)
   - Authorization code (for a "native" service principal)
   - Device code
   - With a username and password (resource owner grant)
-* `get_azure_token` will now cache AAD tokens and refresh them for subsequent sessions. Tokens are cached in a user-specific configuration directory (unlike httr, which saves them in a special file in the R working directory).
+* `get_azure_token` will now cache AAD tokens and refresh them for subsequent sessions. Tokens are cached in a user-specific configuration directory, using the rappdirs package (unlike httr, which saves them in a special file in the R working directory).
 * Token acquisition logic will shortly move to a new package, to allow it to be used by other packages independently of the Resource Manager interface.
 
 ## Other changes